fix cv format, annd add describe in utils/cv
