+++ cmake/lite.cmake
@@ -45,7 +45,7 @@ function (lite_deps TARGET)
     endforeach(var)
     if(LITE_WITH_CV)
       foreach(var ${lite_cv_deps})
-      set(deps ${deps} ${var})
+        set(deps ${deps} ${var})
       endforeach(var)
     endif()
   endif()
