+++ lite/utils/cv/image_flip.h
@@ -13,9 +13,10 @@
 // limitations under the License.
 
 #pragma once
+
 #include <stdint.h>
 #include <vector>
-#include "lite/utils/cv/cv_enum.h"
+#include "lite/utils/cv/paddle_image_preprocess.h"
 namespace paddle {
 namespace lite {
 namespace utils {
