+++ lite/utils/cv/image_rotate.h
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #pragma once
+
 #include <stdint.h>
 #include <vector>
 namespace paddle {
@@ -20,8 +21,6 @@ namespace lite {
 namespace utils {
 namespace cv {
 void rotate_hwc1(
-    const uint8_t* src, uint8_t* dst, int srcw, int srch, float degree);
-void rotate_hwc2(
     const uint8_t* src, uint8_t* dst, int srcw, int srch, float degree);
 void rotate_hwc3(
     const uint8_t* src, uint8_t* dst, int srcw, int srch, float degree);
