+++ lite/utils/cv/paddle_cv_enum.h
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #pragma once
+
 #include <vector>
 namespace paddle {
 namespace lite {
